+++ package.json
@@ -35,6 +35,10 @@
     "@lerna/filter-packages": "^3.13.0",
     "@lerna/project": "^3.13.1",
     "@pixi/jsdoc-template": "^2.5.1",
+    "@rollup/plugin-buble": "^0.21.3",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.1",
+    "@rollup/plugin-typescript": "^8.1.0",
     "acorn": "^6.1.1",
     "copyfiles": "^2.1.0",
     "cross-env": "^5.2.0",
@@ -46,11 +50,8 @@
     "minimist": "^1.2.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^2.6.3",
-    "rollup": "^1.11.3",
-    "rollup-plugin-buble": "^0.19.6",
-    "rollup-plugin-commonjs": "^9.3.4",
-    "rollup-plugin-node-resolve": "^4.2.3",
+    "rollup": "^2.35.1",
     "rollup-plugin-string": "^3.0.0",
-    "rollup-plugin-terser": "^4.0.4"
+    "rollup-plugin-terser": "^7.0.2"
   }
 }
