+++ scripts/get-sorted-packages.js
@@ -1,0 +1,28 @@
+import batchPackages from '@lerna/batch-packages';
+import filterPackages from '@lerna/filter-packages';
+import { getPackages } from '@lerna/project';
+import minimist from 'minimist';
+
+/**
+ * Get a list of the non-private sorted packages with Lerna v3
+ * @see https://github.com/lerna/lerna/issues/1848
+ * @return {Promise<Package[]>} List of packages
+ */
+async function getSortedPackages() {
+    // Support --scope and --ignore globs
+    const {scope, ignore} = minimist(process.argv.slice(2));
+
+    // Standard Lerna plumbing getting packages
+    const packages = await getPackages(__dirname);
+    const filtered = filterPackages(
+        packages,
+        scope,
+        ignore,
+        false
+    );
+
+    return batchPackages(filtered)
+        .reduce((arr, batch) => arr.concat(batch), []);
+}
+
+export default getSortedPackages;
