+++ tools/demo/package.json
@@ -13,14 +13,14 @@
     "cross-env": "^5.2.0",
     "pixi-filters": "3.2.2",
     "rimraf": "^2.6.3",
-    "rollup": "^1.11.3",
-    "rollup-plugin-buble": "^0.19.6",
-    "rollup-plugin-commonjs": "^9.3.4",
+    "rollup": "^2.35.1",
+    "@rollup/plugin-buble": "^0.21.3",
+    "@rollup/plugin-commonjs": "^17.0.0",
+    "@rollup/plugin-node-resolve": "^11.0.1",
+    "@rollup/plugin-typescript": "^8.1.0",
     "rollup-plugin-node-builtins": "^2.1.2",
     "rollup-plugin-node-globals": "^1.4.0",
-    "rollup-plugin-node-resolve": "^4.2.3",
-    "rollup-plugin-terser": "^4.0.4",
-    "rollup-watch": "^4.3.1"
+    "rollup-plugin-terser": "^7.0.2"
   },
   "dependencies": {
     "pixi.js": "^5.0.0"
